requirements.txt
# Add your project dependencies here
openai
dspy-ai
langchain-core
langchain_openai
arxiv
python-dotenv
emoji